Add copyRecursivelySync & copyPreserveSync
Implement linkRecursivelySync & linkAndOverwrite as copy.

We now refuse to overwrite in either of those functions, despite the
`linkAndOverwrite` name. This is slight breakage on a non-major version,
but to my knowledge nobody relies on the overwriting behavior anymore
(thanks GitHub search). Due to the possibility of data loss with hard
links, it seems more important to get the non-linking behavior out in a
minor version, so that plugins using a floating version will pick it up.